add everything except data
