+++ .gitignore
@@ -1,3 +1,8 @@
+# UniProt XML gz files
+*.xml.gz
+*.tsv
+*.tsv.gz
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
